Добавил файл заголовка.
Вынес функции в отдельный модуль.
